src/misc/haystack/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
astor==0.8.1
asttokens==2.4.1
attrs==23.1.0
blessed==1.20.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
executing==2.0.1
fastapi==0.105.0
filelock==3.13.1
FLAML==2.1.1
frozenlist==1.4.1
fsspec==2023.12.2
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
html2image==2.0.4.3
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==7.0.0
inquirer==3.1.4
ipykernel==6.27.1
ipython==8.19.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.1
litellm==0.13.2
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.2
open-interpreter==0.1.17
openai==0.28.1
orjson==3.9.10
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
readchar==4.0.5
regex==2023.10.3
requests==2.31.0
rich==13.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
termcolor==2.4.0
tiktoken==0.4.0
tokenizers==0.15.0
tokentrim==0.1.13
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.12
websocket-client==1.7.0
websockets==12.0
wget==3.2
yarl==1.9.4
yaspin==3.0.1
zipp==3.17.0
